requirements.txt
--index-url https://download.pytorch.org/whl/cu126
--extra-index-url https://pypi.org/simple

# Core dependencies
torch==2.7.1
torchvision==0.22.1
numpy>=1.24.0
pillow>=10.0.0

# Transformers and models
transformers>=4.35.0
accelerate>=0.25.0

# Data processing
webdataset>=0.2.86
scipy==1.15.3
scikit-learn>=1.3.0

# Configuration and logging
hydra-core==1.3.2

# Visualization
matplotlib==3.10.6
opencv-python==4.12.0.88

# Object detection and segmentation
pycocotools==2.0.10

# Utilities
einops
tqdm
py3_wget
requests>=2.31.0

# Jupyter notebook support
jupyter
jupyter_contrib_nbextensions==0.7.0
ipympl==0.9.7
